优化代码结构，继续细分模块
