+++ nonebot_plugin_chatglm6b/request.py
@@ -1,0 +1,41 @@
+from nonebot.log import logger
+
+import aiohttp
+
+from .config import config
+
+class Check:
+    async def chk_server(self):
+        async with aiohttp.ClientSession() as session:
+            async with session.get(f"{config.chatglm_addr}/") as test:
+                if (await test.json())["message"]!="Hello ChatGLM API!":
+                    return False
+                return True
+
+check = Check()
+
+class Request:
+    async def get_resp(txt, history):
+    #res = requests.post(f"{config.chatglm_addr}/predict?user_msg={txt}", json=history)
+        try:
+            async with aiohttp.ClientSession() as session:  #发送请求
+                async with session.post(f"{config.chatglm_addr}/predict?user_msg={txt}", json=history) as res:
+                    if res.status not in [200, 201]:
+                        logger.error(await res.text())
+                        raise RuntimeError(f"与服务器沟通时发生{res.status}错误")
+                    resp, history = (await res.json())["response"], (await res.json())["history"]
+                    return resp, history
+
+        except aiohttp.ServerTimeoutError:  #响应超时
+            logger.error("请求超时。\n" + res)
+            raise RuntimeError(f"可恶，这个AI没反应了，要不炖了吧？")
+
+        except aiohttp.InvalidURL:  #地址错误
+            logger.error("API服务器地址格式有误")
+            raise RuntimeError(f"配置有误，请反馈给我的主人。")
+
+        except Exception as e:  #其他情况
+            logger.error("error:" + str(e) + "\nresponse:" + await res.text())
+            raise RuntimeError(f"请求时出现未知错误：{str(e)}")
+
+request = Request()
